pyproject.toml
[tool.poetry]
name = "thameswaterclient"
version = "0.1.0"
description = "A Python client for retrieving hourly meter data from Thames Water"
authors = ["Ayrton Bourn <ayrtonbourn@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "^2.31.0"
pydantic = "^2.10.6"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
